confirmation to delete user

+++ frontend/src/screens/UserListScreen.js
@@ -30,7 +30,10 @@ function UserListScreen() {
   }, [dispatch, navigate, successDelete]);
 
   const deleteHandler = (id) => {
-    dispatch(deleteUser(id))
+
+    if(window.confirm('Are you sure you want to delete this user?')){
+      dispatch(deleteUser(id))
+    }
   }
 
   return (
